fixing pricing CTA stg->prod issue (#205)
