+++ express/blocks/puf/puf.js
@@ -12,78 +12,11 @@
 import {
   addPublishDependencies,
   createTag,
-  getHelixEnv,
   // eslint-disable-next-line import/no-unresolved
 } from '../../scripts/scripts.js';
-import { getOffer } from '../../scripts/utils/pricing.js';
+import { getOffer, buildUrl } from '../../scripts/utils/pricing.js';
 
 import { buildCarousel } from '../shared/carousel.js';
-
-function replaceUrlParam(url, paramName, paramValue) {
-  const params = url.searchParams;
-  params.set(paramName, paramValue);
-  url.search = params.toString();
-  return url;
-}
-
-function buildUrl(optionUrl, country, language) {
-  const currentUrl = new URL(window.location.href);
-  let planUrl = new URL(optionUrl);
-
-  if (!planUrl.hostname.includes('commerce')) {
-    return planUrl.href;
-  }
-  planUrl = replaceUrlParam(planUrl, 'co', country);
-  planUrl = replaceUrlParam(planUrl, 'lang', language);
-  let rUrl = planUrl.searchParams.get('rUrl');
-  if (currentUrl.searchParams.has('host')) {
-    const hostParam = currentUrl.searchParams.get('host');
-    if (hostParam === 'express.adobe.com') {
-      planUrl.hostname = 'commerce.adobe.com';
-      if (rUrl) rUrl = rUrl.replace('express.adobe.com', hostParam);
-    } else if (hostParam.includes('qa.adobeprojectm.com')) {
-      planUrl.hostname = 'commerce.adobe.com';
-      if (rUrl) rUrl = rUrl.replace('express.adobe.com', hostParam);
-    } else if (hostParam.includes('.adobeprojectm.com')) {
-      planUrl.hostname = 'commerce-stg.adobe.com';
-      if (rUrl) rUrl = rUrl.replace('adminconsole.adobe.com', 'stage.adminconsole.adobe.com');
-      if (rUrl) rUrl = rUrl.replace('express.adobe.com', hostParam);
-    }
-  }
-
-  const env = getHelixEnv();
-  if (env && env.commerce && planUrl.hostname.includes('commerce')) planUrl.hostname = env.commerce;
-  if (env && env.spark && rUrl) {
-    const url = new URL(rUrl);
-    url.hostname = env.spark;
-    rUrl = url.toString();
-  }
-
-  if (rUrl) {
-    rUrl = new URL(rUrl);
-
-    if (currentUrl.searchParams.has('touchpointName')) {
-      rUrl = replaceUrlParam(rUrl, 'touchpointName', currentUrl.searchParams.get('touchpointName'));
-    }
-    if (currentUrl.searchParams.has('destinationUrl')) {
-      rUrl = replaceUrlParam(rUrl, 'destinationUrl', currentUrl.searchParams.get('destinationUrl'));
-    }
-    if (currentUrl.searchParams.has('srcUrl')) {
-      rUrl = replaceUrlParam(rUrl, 'srcUrl', currentUrl.searchParams.get('srcUrl'));
-    }
-  }
-
-  if (currentUrl.searchParams.has('code')) {
-    planUrl.searchParams.set('code', currentUrl.searchParams.get('code'));
-  }
-
-  if (currentUrl.searchParams.get('rUrl')) {
-    rUrl = currentUrl.searchParams.get('rUrl');
-  }
-
-  if (rUrl) planUrl.searchParams.set('rUrl', rUrl.toString());
-  return planUrl.href;
-}
 
 function pushPricingAnalytics(adobeEventName, sparkEventName, plan) {
   const url = new URL(window.location.href);
